requirements.txt
streamlit
langchain
langchain-community
langchain-groq
faiss-cpu
plotly
pypdf
python-dotenv
